database: Moved benchmarks to `benchmark_database.go`.
